requirements.txt
Django==1.8.2
Django-Select2==4.3.1
Pillow==2.8.2
argparse==1.2.1
dj-database-url==0.3.0
dj-static==0.0.6
django-bootstrap-pagination==1.5.1
django-markdown-deux==1.0.5
django-toolbelt==0.0.1
gunicorn==19.3.0
markdown2==2.3.0
psycopg2==2.6.1
simplejson==3.7.3
sorl-thumbnail==12.2
static3==0.6.1
wsgiref==0.1.2
